requirements.txt
streamlit
python-docx
# fpdf4
PyPDF2
transformers
torch
docx2pdf